Finally get rid of Optionable entirely.
Also convert its old test to new-style.

[ci skip-rust]

[ci skip-build-wheels]
